+++ src/App.ts
@@ -7,13 +7,7 @@ import { bindings } from "./inversify.config";
 import express from "express";
 import { CustomAuthProvider } from "./apiBase/auth";
 import cors from "cors"
-import { ApolloServer } from 'apollo-server-express'
-import depthLimit from 'graphql-depth-limit'
-import { importSchema } from 'graphql-import';
-import resolvers from './graphql/resolvers'
-import { ReqContext } from './graphql/types/server.types';
-import { peopleFromHouseholdLoader, householdLoader } from './graphql/loader'
-import { validateToken } from './graphql/helpers/account';
+import { GraphQLHelper } from "./graphql/helpers/GraphQLHelper";
 
 export const init = async () => {
   /*
@@ -39,24 +33,8 @@ export const init = async () => {
   };
 
   const server = app.setConfig(configFunction).build();
-  const graphQLServer = new ApolloServer({
-    typeDefs: importSchema('src/graphql/schema/schema.graphql'),
-    resolvers,
-    validationRules: [depthLimit(5)],
-    context: (ctx: ReqContext) => {
-      const { authorization } = ctx.req.headers
-      if (authorization && typeof authorization === 'string' && authorization.startsWith('Bearer ')) {
-        const token = authorization.substr(7)
-        const me = validateToken(token)
-        ctx.me = me
-      }
-      ctx.peopleFromHouseHoldLoader = peopleFromHouseholdLoader()
-      ctx.householdLoader = householdLoader()
 
-      return ctx
-    },
-  })
-
+  const graphQLServer = GraphQLHelper.getServer();
   graphQLServer.applyMiddleware({
     app: server,
     path: '/graphql',
