+++ src/graphql/types/server.types.ts
@@ -1,11 +1,11 @@
 import { ExpressContext } from 'apollo-server-express';
 import { AuthenticatedUser } from '../../apiBase/auth';
 import { IMe } from '../helpers/account';
-import { HouseholdLoader, PeopleFromHouseholdLoader } from '../loader';
+import { HouseholdLoaderType, PeopleFromHouseholdLoaderType } from '../loader';
 
 export type ReqContext = ExpressContext & {
   me?: IMe
   au?: AuthenticatedUser
-  peopleFromHouseHoldLoader: PeopleFromHouseholdLoader
-  householdLoader: HouseholdLoader
+  peopleFromHouseHoldLoader: PeopleFromHouseholdLoaderType
+  householdLoader: HouseholdLoaderType
 }
